pyproject.toml
[tool.poetry]
name = "lsu-pylot"
version = "0.1.0"
description = ""
authors = ["Darko Kolev <kolevdarko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.6"
openai = "^1.52.1"
python-dotenv = "^1.0.1"
tiktoken = "^0.8.0"
pandas = "^2.2.3"
langchain = "^0.3.4"
scipy = "^1.14.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
